Added separate attest executable
